requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
opencv-python==4.10.0.84
ultralytics==8.3.0
numpy==1.26.4
Pillow==10.4.0
torch==2.5.1
torchvision==0.20.1
sumo-rl==1.4.5
gymnasium==0.29.1
pandas==2.2.3
matplotlib==3.9.2
fire==0.7.1
